requirements.txt
-i https://pypi.org/simple
altair==5.5.0; python_version >= '3.9'
annotated-types==0.7.0; python_version >= '3.8'
anyio==4.9.0; python_version >= '3.9'
attrs==25.3.0; python_version >= '3.8'
beautifulsoup4==4.13.3; python_full_version >= '3.7.0'
blinker==1.9.0; python_version >= '3.9'
cachetools==5.5.2; python_version >= '3.7'
certifi==2025.1.31; python_version >= '3.6'
charset-normalizer==3.4.1; python_version >= '3.7'
click==8.1.8; python_version >= '3.7'
cloudevents==1.11.0
deprecation==2.1.0
dnspython==2.7.0; python_version >= '3.9'
docstring-parser==0.16; python_version >= '3.6' and python_version < '4.0'
email-validator==2.2.0; python_version >= '3.8'
fastapi[standard]==0.115.12; python_version >= '3.8'
fastapi-cli[standard]==0.0.7; python_version >= '3.8'
flask==3.1.0; python_version >= '3.9'
functions-framework==3.8.2; python_version >= '3.5' and python_version < '4'
gitdb==4.0.12; python_version >= '3.7'
gitpython==3.1.44; python_version >= '3.7'
google-api-core[grpc]==2.24.2; python_version >= '3.7'
google-auth==2.38.0; python_version >= '3.7'
google-cloud-aiplatform==1.86.0; python_version >= '3.8'
google-cloud-bigquery==3.31.0; python_version >= '3.9'
google-cloud-core==2.4.3; python_version >= '3.7'
google-cloud-resource-manager==1.14.2; python_version >= '3.7'
google-cloud-storage==2.19.0; python_version >= '3.7'
google-crc32c==1.7.1; python_version >= '3.9'
google-genai==1.8.0; python_version >= '3.9'
google-resumable-media==2.7.2; python_version >= '3.7'
googleapis-common-protos[grpc]==1.69.2; python_version >= '3.7'
grpc-google-iam-v1==0.14.2; python_version >= '3.7'
grpcio==1.71.0; python_version >= '3.9'
grpcio-status==1.71.0; python_version >= '3.9'
gunicorn==23.0.0; python_version >= '3.7'
h11==0.14.0; python_version >= '3.7'
httpcore==1.0.7; python_version >= '3.8'
httptools==0.6.4; python_full_version >= '3.8.0'
httpx==0.28.1; python_version >= '3.8'
idna==3.10; python_version >= '3.6'
itsdangerous==2.2.0; python_version >= '3.8'
jinja2==3.1.6; python_version >= '3.7'
jsonschema==4.23.0; python_version >= '3.8'
jsonschema-specifications==2024.10.1; python_version >= '3.9'
markdown-it-py==3.0.0; python_version >= '3.8'
markupsafe==3.0.2; python_version >= '3.9'
mdurl==0.1.2; python_version >= '3.7'
narwhals==1.32.0; python_version >= '3.8'
numpy==2.2.4; python_version >= '3.10'
packaging==24.2; python_version >= '3.8'
pandas==2.2.3; python_version >= '3.9'
pillow==11.1.0; python_version >= '3.9'
proto-plus==1.26.1; python_version >= '3.7'
protobuf==5.29.4; python_version >= '3.8'
pyarrow==19.0.1; python_version >= '3.9'
pyasn1==0.6.1; python_version >= '3.8'
pyasn1-modules==0.4.2; python_version >= '3.8'
pydantic==2.11.1; python_version >= '3.9'
pydantic-core==2.33.0; python_version >= '3.9'
pydeck==0.9.1; python_version >= '3.8'
pygments==2.19.1; python_version >= '3.8'
pypdf2==3.0.1; python_version >= '3.6'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
python-dotenv==1.1.0; python_version >= '3.9'
python-multipart==0.0.20; python_version >= '3.8'
pytz==2025.2
pyyaml==6.0.2; python_version >= '3.8'
referencing==0.36.2; python_version >= '3.9'
requests==2.32.3; python_version >= '3.8'
rich==14.0.0; python_full_version >= '3.8.0'
rich-toolkit==0.14.1; python_version >= '3.8'
rpds-py==0.24.0; python_version >= '3.9'
rsa==4.9; python_version >= '3.6' and python_version < '4'
shapely==2.0.7; python_version >= '3.7'
shellingham==1.5.4; python_version >= '3.7'
six==1.17.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
smmap==5.0.2; python_version >= '3.7'
sniffio==1.3.1; python_version >= '3.7'
soupsieve==2.6; python_version >= '3.8'
starlette==0.46.1; python_version >= '3.9'
streamlit==1.44.0; python_version >= '3.9' and python_full_version != '3.9.7'
tenacity==9.0.0; python_version >= '3.8'
toml==0.10.2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2'
tornado==6.4.2; python_version >= '3.8'
typer==0.15.2; python_version >= '3.7'
typing-extensions==4.13.0; python_version >= '3.8'
typing-inspection==0.4.0; python_version >= '3.9'
tzdata==2025.2; python_version >= '2'
urllib3==2.3.0; python_version >= '3.9'
uvicorn[standard]==0.34.0; python_version >= '3.9'
uvloop==0.21.0; python_full_version >= '3.8.0'
watchdog==6.0.0; python_version >= '3.9'
watchfiles==1.0.4; python_version >= '3.9'
websockets==15.0.1; python_version >= '3.9'
werkzeug==3.1.3; python_version >= '3.9'
